# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bff_elearning",
    "version": "1.0.0"
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "BAD_REQUEST",
                  "UNAUTHORIZED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "CONFLICT",
                  "PAYLOAD_TOO_LARGE",
                  "UNPROCESSABLE_ENTITY",
                  "TOO_MANY_REQUESTS",
                  "INTERNAL_ERROR",
                  "BAD_GATEWAY",
                  "SERVICE_UNAVAILABLE",
                  "GATEWAY_TIMEOUT"
                ]
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            },
            "required": [
              "code",
              "message",
              "details"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/elearning/profile": {
      "get": {
        "tags": [
          "E-learning"
        ],
        "summary": "Loads the E-learning profile",
        "description": "The caller as resolved by BFF User `/me`, ready to render on the profile page.",
        "responses": {
          "200": {
            "description": "Profile loaded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "user-123"
                        },
                        "name": {
                          "type": "string",
                          "example": "Admin Système"
                        },
                        "initials": {
                          "type": "string",
                          "example": "AS"
                        },
                        "email": {
                          "type": "string",
                          "format": "email",
                          "example": "admin@mairie360.fr"
                        },
                        "phone": {
                          "type": "string",
                          "example": "+262 692 00 00 00"
                        },
                        "service": {
                          "type": "string",
                          "example": "Administration"
                        },
                        "position": {
                          "type": "string",
                          "example": "Administrateur système"
                        },
                        "role": {
                          "type": "string",
                          "example": "admin"
                        },
                        "isAdmin": {
                          "type": "boolean",
                          "example": true
                        },
                        "avatarUrl": {
                          "type": "string",
                          "format": "uri",
                          "example": "https://example.com/avatar.jpg"
                        },
                        "address": {
                          "type": "string",
                          "example": "1 rue de la Mairie"
                        },
                        "city": {
                          "type": "string",
                          "example": "Saint-Denis"
                        },
                        "lastConnection": {
                          "type": "string",
                          "example": "3 juillet 2026 à 09:15"
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "initials",
                        "isAdmin"
                      ],
                      "description": "Utilisateur connecté tel qu’attendu par les écrans E-learning"
                    },
                    "footer": {
                      "type": "object",
                      "properties": {
                        "productName": {
                          "type": "string",
                          "example": "Mairie360"
                        },
                        "version": {
                          "type": "string",
                          "example": "2.1.0"
                        },
                        "links": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "label": {
                                "type": "string",
                                "example": "Support technique"
                              },
                              "href": {
                                "type": "string",
                                "example": "/support"
                              }
                            },
                            "required": [
                              "label",
                              "href"
                            ],
                            "description": "Lien affiché dans le footer"
                          }
                        }
                      },
                      "required": [
                        "productName",
                        "version",
                        "links"
                      ],
                      "description": "Configuration du footer"
                    }
                  },
                  "required": [
                    "user"
                  ],
                  "description": "Réponse de chargement de la page profil"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "BFF User or an upstream API is unreachable, not configured, failed or answered an unusable body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "E-learning"
        ],
        "summary": "Updates the editable profile fields",
        "description": "`email` and `phone` are saved by Core API (`PATCH /api/v1/user/me/`), then the profile is read again from BFF User. `address` and `city` are stored by no upstream service: a body carrying them answers 501 and nothing is written. Non-editable fields (role, isAdmin, service, position, lastConnection) are ignored.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "security-admin@mairie360.fr"
                  },
                  "phone": {
                    "type": "string",
                    "example": "0612345678"
                  },
                  "address": {
                    "type": "string",
                    "example": "1 place de la Mairie"
                  },
                  "city": {
                    "type": "string",
                    "example": "Paris"
                  }
                },
                "description": "Champs éditables du profil utilisateur"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile saved, then read again",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "user-123"
                        },
                        "name": {
                          "type": "string",
                          "example": "Admin Système"
                        },
                        "initials": {
                          "type": "string",
                          "example": "AS"
                        },
                        "email": {
                          "type": "string",
                          "format": "email",
                          "example": "admin@mairie360.fr"
                        },
                        "phone": {
                          "type": "string",
                          "example": "+262 692 00 00 00"
                        },
                        "service": {
                          "type": "string",
                          "example": "Administration"
                        },
                        "position": {
                          "type": "string",
                          "example": "Administrateur système"
                        },
                        "role": {
                          "type": "string",
                          "example": "admin"
                        },
                        "isAdmin": {
                          "type": "boolean",
                          "example": true
                        },
                        "avatarUrl": {
                          "type": "string",
                          "format": "uri",
                          "example": "https://example.com/avatar.jpg"
                        },
                        "address": {
                          "type": "string",
                          "example": "1 rue de la Mairie"
                        },
                        "city": {
                          "type": "string",
                          "example": "Saint-Denis"
                        },
                        "lastConnection": {
                          "type": "string",
                          "example": "3 juillet 2026 à 09:15"
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "initials",
                        "isAdmin"
                      ],
                      "description": "Utilisateur connecté tel qu’attendu par les écrans E-learning"
                    }
                  },
                  "required": [
                    "user"
                  ],
                  "description": "Profil mis à jour"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload, or a value Core API refuses (e.g. a phone number that is not 10 to 15 digits)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "The e-mail address is already used by another account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not available: no upstream service stores this data yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "BFF User or an upstream API is unreachable, not configured, failed or answered an unusable body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}